Add files via upload
Background of the problem statement:

The company has created a new website for the LiveDocApp where the patient can add their details in the website. The company wants to use a public cloud for the internet facing website of LiveDocApp. Once the website was deployed, patients started complaining that the pages are not loading fast enough. You realized that the website gets global traffic and the static assets like pages are served from a single server. You need to make sure that the traffic coming to the application from different parts of the world is load balanced at the DNS level.

You can use either Azure or AWS platforms to design the solution using IaaS OR PaaS.